Add Navbar and Fix Folder Structure
