+++ client/src/pages/trainer/course/detail/[id].tsx
